+++ lib/helper.rb
@@ -62,6 +62,18 @@ module Creq
       }
     end
 
+    def create_pandoc(format, options)
+      # TODO check is pandoc available
+      req = inside_req { Repository.load }
+      inside_doc {
+        open("pandoctmp.md", 'w') {|f|
+          PandocWriter.write(req, f)
+          # pandoc -s --reference-doc template.docx --toc -o srs.docx srs.md
+          `pandoc -s --toc #{options} -o requirements.#{format} pandoctmp.md`
+        }
+      }
+    end
+
     def check_repo
       repo = inside_req { Repository.load }
       {}.tap do |err|